+++ src/HttpClient/context.ts
@@ -1,6 +1,8 @@
 import {AxiosRequestConfig, AxiosResponse} from 'axios'
+import {IAxiosRetryConfig} from 'axios-retry'
 
 export interface RequestConfig extends AxiosRequestConfig {
+  'axios-retry'?: IAxiosRetryConfig
   metric?: string
   production?: boolean
 }