+++ src/HttpClient/index.ts
@@ -1,8 +1,9 @@
 import { DataSource } from 'apollo-datasource'
 import { AxiosResponse } from 'axios'
+import { IAxiosRetryConfig } from 'axios-retry'
 import { IncomingMessage } from 'http'
 import { Context } from 'koa'
-import * as compose from 'koa-compose'
+import compose from 'koa-compose'
 import { ParsedUrlQuery } from 'querystring'
 
 import { CacheLayer } from '../caches/CacheLayer'
@@ -43,26 +44,26 @@ export class HttpClient {
 
   public static forWorkspace (service: string, context: IOContext, opts: InstanceOptions): HttpClient {
     const {authToken, userAgent, recorder, segmentToken, sessionToken} = context
-    const {timeout, cacheStorage} = opts
+    const {timeout, cacheStorage, retryConfig} = opts
     const baseURL = workspaceURL(service, context, opts)
-    return new HttpClient({baseURL, authType: AuthType.bearer, authToken, userAgent, timeout, recorder, cacheStorage, segmentToken, sessionToken})
+    return new HttpClient({baseURL, authType: AuthType.bearer, authToken, userAgent, timeout, recorder, cacheStorage, segmentToken, sessionToken, retryConfig})
   }
 
   public static forRoot (service: string, context: IOContext, opts: InstanceOptions): HttpClient {
     const {authToken, userAgent, recorder, segmentToken, sessionToken} = context
-    const {timeout, cacheStorage} = opts
+    const {timeout, cacheStorage, retryConfig} = opts
     const baseURL = rootURL(service, context, opts)
-    return new HttpClient({baseURL, authType: AuthType.bearer, authToken, userAgent, timeout, recorder, cacheStorage, segmentToken, sessionToken})
+    return new HttpClient({baseURL, authType: AuthType.bearer, authToken, userAgent, timeout, recorder, cacheStorage, segmentToken, sessionToken, retryConfig})
   }
 
   public static forLegacy (endpoint: string, opts: LegacyInstanceOptions): HttpClient {
-    const {authToken, userAgent, timeout, cacheStorage} = opts
-    return new HttpClient({baseURL: endpoint, authType: AuthType.token, authToken, userAgent, timeout, cacheStorage})
+    const {authToken, userAgent, timeout, cacheStorage, retryConfig} = opts
+    return new HttpClient({baseURL: endpoint, authType: AuthType.token, authToken, userAgent, timeout, cacheStorage, retryConfig})
   }
   private runMiddlewares: compose.ComposedMiddleware<MiddlewareContext>
 
   public constructor (opts: ClientOptions) {
-    const {baseURL, authToken, authType, cacheStorage, metrics, recorder, userAgent, timeout = DEFAULT_TIMEOUT_MS, segmentToken} = opts
+    const {baseURL, authToken, authType, cacheStorage, metrics, recorder, userAgent, timeout = DEFAULT_TIMEOUT_MS, segmentToken, retryConfig} = opts
     const headers: Record<string, string> = {
       'Accept-Encoding': 'gzip',
       'User-Agent': userAgent,
@@ -73,7 +74,7 @@ export class HttpClient {
     }
 
     this.runMiddlewares = compose([
-      defaultsMiddleware(baseURL, headers, timeout),
+      defaultsMiddleware(baseURL, headers, timeout, retryConfig),
       ...recorder ? [recorderMiddleware(recorder)] : [],
       acceptNotFoundMiddleware,
       ...cacheStorage ? [cacheMiddleware({cacheStorage, segmentToken: segmentToken || ''})] : [],
@@ -174,6 +175,7 @@ export interface InstanceOptions {
   timeout?: number,
   cacheStorage?: CacheLayer<string, Cached>,
   endpoint?: string,
+  retryConfig?: IAxiosRetryConfig,
 }
 
 export interface LegacyInstanceOptions {
@@ -182,6 +184,7 @@ export interface LegacyInstanceOptions {
   timeout?: number,
   accept?: string,
   cacheStorage?: CacheLayer<string, Cached>,
+  retryConfig?: IAxiosRetryConfig,
 }
 
 export interface IOResponse<T> {
@@ -206,4 +209,5 @@ interface ClientOptions {
   cacheStorage?: CacheLayer<string, Cached>,
   segmentToken?: string
   sessionToken?: string
+  retryConfig?: IAxiosRetryConfig
 }