+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.1.0] - 2019-03-07
+
+### Changed
+- Change default retry behaviour to only retry on network error once
+- Accept retryConfig on each HttpClient
+
+## [2.1.0-beta.1] - 2019-02-28
+
+## [2.1.0-beta.0] - 2019-02-27
+
+## [2.1.0-beta] - 2019-02-27
 ### Changed
 - Added new required parameter to create a new workspace on Workspaces API.